Fix types, bounds, and rail switch logic
Improve robustness and prevent UB/infinite loops across train model code. Key changes:

- Replace char 'N' with numeric 0 for switch placeholders and planned switches initialization.
- Change idle timing and loop-timing variables from int to unsigned long to match millis() and avoid overflow.
- Add maxTries/tries guards to set_target_rail_section_randomly and set_target_rail_section_by_queue to avoid infinite loops; increment tries and only advance to pathfinding when a valid station-containing target is found. Add debug Serial messages when no valid target is found.
- Tighten update_rail_switches: derive previous section from plannedRailSectionsNames[0], check planned switch index != -1 before accessing arrays, and only call set_switch when a change is needed (prevents out-of-bounds/invalid accesses).
- Shrink Proximity_IR and Rail_Switch_DC initializers to match actual configured pins/hardware (removed extra initializers).

These changes are aimed at preventing crashes, infinite loops, and incorrect switch operations in edge cases and on long-running timers.

+++ Train_DC.h
@@ -36,8 +36,8 @@ class Train_DC
         bool _switchRealPositiveDirectionStatus = false;
         bool _switchDirectionStatus = false;
         // Attributes - Idle
-        int _idleStartTime_ms;
-        int _currentIdleTime;
+        unsigned long _idleStartTime_ms;
+        unsigned long _currentIdleTime;
         // Attributes - Rail Exploration
         int _realPositiveDirection = -1; // Real Direction when Reverser is Positive (1). 1 is Forward, -1 is Backward. Modulates reverser.
         // Attributes - Rail Localization - Current & Target (Must Reflect Present Truth)
